it's kinda running, at about the same performance/functionality as before
