+++ pixel_engine/StagingHost.cs
@@ -1,13 +1,53 @@
 ﻿
+using System.Collections.Generic;
 using System.Numerics;
-using System.Security.Policy;
-using Point = System.Windows.Point;
 
 namespace pixel_core
 {
     public class StagingHost
     {
         public Node? lastSelected;
+        public static Stage Standard()
+        {
+            var nodes = new List<Node>();
+
+            Node camera = new("camera");
+            camera.AddComponent<Camera>();
+            camera.Position = new(0, 0);
+
+            Node floor = Floor.Standard();
+            nodes.Add(floor);
+
+            Node textTest = new("text");
+            var img = textTest.AddComponent<Text>();
+
+            //TODO: Eliminate this confusion, I don't know which one's even in use atm.
+            img.viewportSize = new(25, 25);
+            textTest.Scale = new(25, 25);
+            img.viewportPosition = Vector2.Zero;
+            textTest.Position = Vector2.Zero;
+
+
+            Node light = new("light");
+            var lt = light.AddComponent<Light>();
+            light.Position = new(50, -50);
+            lt.brightness = 0.9f;
+            lt.radius = 10;
+
+            nodes.Add(textTest);
+            nodes.Add(light);
+            nodes.Add(camera);
+
+            for (int i = 0; i < 5; i++)
+            {
+                Node rbNode = Rigidbody.Standard();
+                rbNode.Position = new(i * 20, -20);
+                nodes.Add(rbNode);
+            }
+
+            var stage = new Stage("default stage", Stage.DefaultBackgroundMetadata, nodes);
+            return stage;
+        }
         /// <summary>
         /// this is used for editor clicking.
         /// </summary>
